Move IsCert from common into Node
