+++ src/util/common.ts
@@ -1,17 +1,5 @@
 import blake2b from "blake2b"
 
-import {
-    PRIMARY_INTERFACE_CHAINCERT_ID,
-} from "../datamodel/cert/primary/interface/PrimaryChainCertInterface";
-
-import {
-    PRIMARY_INTERFACE_DEFAULTCERT_ID,
-} from "../datamodel/cert/primary/interface/PrimaryDefaultCertInterface";
-
-import {
-    PRIMARY_INTERFACE_NODECERT_ID,
-} from "../datamodel/cert/primary/interface/PrimaryNodeCertInterface";
-
 /**
  * Compares A and B on creationTime and if equal then compare also on ID.
  *
@@ -324,28 +312,4 @@ export function PromiseCallback<ReturnType>(): {promise: Promise<ReturnType>, cb
         promise,
         cb,
     };
-}
-
-/**
-* Check in the image header data if it looks like a cert.
-* @param image the cert image
-* @returns true if the image header is recognized as being of primary interface cert.
-*/
-export function IsCert(image: Buffer): boolean {
-    const chainCertPrimaryInterface = Buffer.from([0, PRIMARY_INTERFACE_CHAINCERT_ID]);
-    if (image.slice(0, 2).equals(chainCertPrimaryInterface)) {
-        return true;
-    }
-
-    const defaultCertPrimaryInterface = Buffer.from([0, PRIMARY_INTERFACE_DEFAULTCERT_ID]);
-    if (image.slice(0, 2).equals(defaultCertPrimaryInterface)) {
-        return true;
-    }
-
-    const nodeCertPrimaryInterface = Buffer.from([0, PRIMARY_INTERFACE_NODECERT_ID]);
-    if (image.slice(0, 2).equals(nodeCertPrimaryInterface)) {
-        return true;
-    }
-
-    return false;
 }
